fix control pid transfer function K_D factor
factor was computed correctly, but documented transfer function had a wrong usage of K_D, because simulation
is performed in 0.001s = 1s to avoid numeric computation issues. The correction of this term was not considered
when copying the transfer function to the documentation. (Altough the correction was considered for the
computation of K_D.)